+++ lib/ui/src/app.stories.js
@@ -4,7 +4,7 @@ import { Root as App } from './index';
 import { PrettyFakeProvider, FakeProvider } from './FakeProvider';
 
 export default {
-  title: 'UI/Layout/App',
+  title: 'UI/App',
   component: App,
 };
