+++ lib/ui/src/components/preview/iframe.stories.js
@@ -4,10 +4,10 @@ import { IFrame } from './iframe';
 
 export default {
   component: IFrame,
-  title: 'UI/Preview/Iframe',
+  title: 'UI/Iframe',
 };
 
-const style = { width: '100%', height: '500px' };
+const style = { width: '500px', height: '500px', border: '2px solid hotpink' };
 
 export const workingStory = () => (
   <IFrame
@@ -30,21 +30,3 @@ export const missingStory = () => (
     scale={1.0}
   />
 );
-
-export const errorStory = () => (
-  <IFrame
-    id="iframe"
-    title="Missing"
-    src="/iframe.html?id=core-errors--story-throws-exception"
-    allowFullScreen
-    style={style}
-    scale={1.0}
-  />
-);
-// We need to disable this one in Chromatic because the screenshot includes the uploaded URL sadly:
-//   eg. https://www.chromaticqa.com/snapshot?appId=5a375b97f4b14f0020b0cda3&id=5c52edb4323f9000249aae72
-errorStory.story = {
-  parameters: {
-    chromatic: { disable: true },
-  },
-};
