+++ lib/ui/src/components/preview/preview.stories.js
@@ -12,7 +12,7 @@ import { previewProps } from './preview.mockdata';
 const provider = new PrettyFakeProvider();
 
 export default {
-  title: 'UI/Preview/Preview',
+  title: 'UI/Preview',
   component: Preview,
   decorators: [
     StoryFn => (
